`parse_fw_string`: hold the firmware reply parser once, in `protocol/text/framing`
`parse_star_fw_string` exists twice: once in `hamilton/transport/usb/protocol.py`, once in
`legacy/liquid_handling/backends/hamilton/STAR_backend.py`. The two are identical - 117 lines each,
no differing lines ignoring comments and docstrings - so a fix to the reply grammar has to be made
in both, and a fix to one is silently not a fix to the other.

The grammar is also not a property of USB. The same shape comes back from a STAR over USB and from
a tilt module over a serial line, and the parser is a pure function of two strings: no state, no
I/O, no device. Keeping it inside the transport that happens to carry it is what let the second
copy appear.

- Adds `protocol/text/framing.py` holding `parse_fw_string` and `parse_firmware_version_date`,
  moved verbatim.
- `transport/usb/protocol.py` re-exports both under their old names, so `parse_star_fw_string` and
  `parse_star_firmware_version_date` keep working from where they are imported today.
- `STAR_backend.py` drops its copy and imports the one implementation, still exporting
  `parse_star_fw_string` for its own callers and tests.

Behaviour: none. The implementation is unchanged, both old names resolve to the same object as the
new ones, and nothing that imports them needs to move.
